requirements.txt
langchain==0.3.0
langchain-community==0.3.0
chromadb==0.5.5
openai==1.41.0
python-dotenv==1.0.1
pypdf==4.2.0
python-pptx==0.6.23
pandas==2.2.2
openpyxl==3.1.5
fastapi==0.115.0
uvicorn[standard]==0.30.5
